+++ changelog.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking
+- remove deprecated methods from `LnurlAuthWalletToken`
+
+
 ## [0.13.0] - 2024-02-14
 
 ### Breaking